Валидации вынесены в отдельные классы Вся логика убрана из контроллера Апдейт цели теперь PUT и предполагает наличие всех полей в запросе Эксепшн хэндлеры вынесены в ControllerAdvice
