+++ src/main/java/school/faang/user_service/validator/goal/GoalValidationParams.java
@@ -1,12 +1,12 @@
 package school.faang.user_service.validator.goal;
 
 import school.faang.user_service.dto.goal.GoalDto;
-import school.faang.user_service.util.Util;
+import school.faang.user_service.util.Helper;
 
 import java.io.Serializable;
 
 public record GoalValidationParams(Long userId, GoalDto goalRq, String path) implements Serializable {
     public String path(String path) {
-        return Util.createPath(this.path, path);
+        return Helper.createPath(this.path, path);
     }
 }
